Save/restore regs when reading/writing memory. Add SFR access.
The DPL and ACC registers are used by the memory access code,
so they need to be saved and restored. Stuff them up high in ram for now;
this should probably be fixed to pull them back to the host instead.

Special SFR access is required as not all SFRs are visible in the unified
address space.

Signed-off-by: Keith Packard <keithp@keithp.com>