Wait for a while when switching the RESET_N line
The cc1111 manual suggests placing a 2.7kΩ resister and 1nF capacitor on the
RESET_N line to filter out noise. This increases the time necessary to reset
the chip to several microseconds which is longer than the interval between
two USB packets. Flush the USB packet queue and sleep for a while after
changing the value on the RESET_N line to make sure the chip sees the state
change.

Signed-off-by: Keith Packard <keithp@keithp.com>