Enabling apogee detect via speed: < 200m/s && < max_speed - 50m/s
This change ensures that we actually got going fairly fast, and then slowed
down a bunch before enabling apogee detect. Otherwise, we'll detect apogee
right off the pad as we're not going very fast at that point...

This also adds the 'f' command to show the current flight status on the USB
port.