Discard usb output before connection. Handle USB reset.
Discarding output before USB is plugged in allows threads that send output
and do other things to work without a USB connection. Unfortuantely, there
doesn't appear to be any way to detect when the USB link is disconnected,
which means that once USB is enabled, future writes will continue to block.

USB reset causes the USB interrupts to all be reconfigured back to power-on
state.

Signed-off-by: Keith Packard <keithp@keithp.com>