Remove bit-banging debug support from TM board builds
Our current TM boards don't have the wires to do bit-banging to
another cc1111 board, so it doesn't make sense to fill up their flash
with useless code (and the 'help' text with useless commands). Leave
this to the TI board until we have boards that can serve as debug dongles.

Signed-off-by: Keith Packard <keithp@keithp.com>