Set all of the values to reset the radio for telemetry
Was sizeof(rdf_setup) instead of sizeof(telemetry_setup) when
resetting the radio back to telemetry data mode from rdf mode. With
the length value removed from the rdf config, these two arrays are no
longer the same length, and so the last config value was not set
leaving the radio sending garbage.

Signed-off-by: Keith Packard <keithp@keithp.com>