Loosen tolerances for main->landed transition
Detecting that the rocket has landed is required for the system to
flush the eeprom log and re-enable the RDF beacon. This patch changes
the landed state entry requirements for the accelerometer to require
only that the accelerometer stay within a quarter of a g (down from
1/10g) and changes the testing interval from 20 seconds to 5
seconds.

The requirement that the barometric altitude be within 1000m of the
launch altitude and that the barometer change by no more than 0.05kPa
are unchanged.

Signed-off-by: Keith Packard <keithp@keithp.com>