Automatically extract flight number for eeprom and telem filenames.
Extract flight number from either telemetry or eeprom files and use
that in the resulting filenames. To ensure that files remain unique,
add a new field, -seq-%03d. This is appended only when the sequence
number is non-zero as it shouldn't occur in normal usage.

This also eliminates some duplicate filename creation code in the
library and aoview sources.

Signed-off-by: Keith Packard <keithp@keithp.com>