Round radio calibration value instead of truncating
The radio calibration function in the bring-up code was truncating the
radio calibration value which caused the resulting frequency to
always measure above the target frequency, instead of trying to get as
close as possible. This change will result in a closer frequency
match, but may sometimes be below the target frequency.

Signed-off-by: Keith Packard <keithp@keithp.com>