Only have the slave return a packet if it received one.
When the receive is aborted to switch modes, it's important to not
immediately re-acquire the radio and try to send a packet as the
aborting thread won't know to kick the receiver again.

This prevents the 'C' command from locking up as it tries to stop the
packet slave before turning on the transmitter.

Signed-off-by: Keith Packard <keithp@keithp.com>