Disable interrupts while reading from stdin
With multiple input source support, there is a lag between asking a
device if it has data and then waiting for more data to appear. If an
interrupt signalling additional input arrives in this interval, we'll
go to sleep with input available.

This patch uses a big hammer by just disabling interrupts for the
whole process.

Signed-off-by: Keith Packard <keithp@keithp.com>