When the EP0 IN buffer is full, don't panic, just skip sending another
If the host doesn't pull the IN packet out of EP0 before sending
another SETUP command along, the IN buffer will still be busy when we
try to reply to the SETUP command. While I don't quite understand why
this would ever happen, there's no need to panic about it, just drop
the reply packet on the floor.

Signed-off-by: Keith Packard <keithp@keithp.com>