altos: Place rom config variables in fixed location
The device serial number and radio calibration values are stored in
flash, mostly so that TeleDongle gets them saved.

Placing them in well-known locations (starting at 0xa0) makes it
possible to find the previous configuration and to re-write it
easily, without requiring the .map file.

altosui doesn't have the .map file parsing code, so it relies upon
this new technique. As a benefit, it reads the old values from the
device before reprogramming it.

Signed-off-by: Keith Packard <keithp@keithp.com>