altosui: provide separate flush_input/flush_output for serial. deal with monitor automatically
(yes, this should be two patches, but the diffs in AltosSerial were merged together).

First, this replaces the existing flush/flush_reply mess with two simple functions,
one to flush output to the serial device, making sure that all data written will be seen
while we wait for input. The other sucks any pending input off of the serial line and
discards it.

Second, AltosSerial now tracks whether the serial line is being used for telemetry
monitoring. If so, it enables monitoring, otherwise it disables it. Eliminates a
bunch of manual state tracking elsewhere.

Signed-off-by: Keith Packard <keithp@keithp.com>