altos: shut down packet mode cleanly
Instead of constantly bashing the packet master thread, let it shut
itself down in an orderly fashion. It will shut down fairly quickly as
all of the activities in that thread are bounded. Otherwise, the
master packet thread might leave mutexes locked and all sorts of other
horrors.

Tested on Linux and Mac OS X and shown to be reliable.

Signed-off-by: Keith Packard <keithp@keithp.com>