altosui: Record flight number when scanning file, not when running
The very first record in the eeprom is the flight number, but it is
time-stamped with the 'boost' time, and so it gets sorted until much
later, delaying the return of data until the rocket enters boost
mode. This drops all of the nice pad GPS and state date on the floor.

Signed-off-by: Keith Packard <keithp@keithp.com>