altosui: Stop parsing eeprom file after hitting 'landed' state
Sometimes there are additional records found in the eeprom file; the
reader is mostly worried about not losing anything, so it reads as
much as it can. However, the last record written for any flight is the
'landed' record, so we can stop looking at the file after hitting that.

Signed-off-by: Keith Packard <keithp@keithp.com>