altosui: Deal with eeprom dates going backwards across wrap
eeprom timestamps can go backwards as the GPS time stamps are
recorded when the first GPS character is received, but not placed into
the eeprom log until the last GPS packet is complete. If this happens
at the same time the tick count is wrapping, then the tick count will
wrap backwards across the 0 boundary causing time to jump forwards.

Fix this by letting time go backwards across the tick boundary, which
requires that we know when the first 'real' tick is read from the
eeprom file.

Signed-off-by: Keith Packard <keithp@keithp.com>