altosui: Open serial device at 'new' time. Prohibit duplicate opens.
With the per-serial UI, there's never a reason to create a serial
device without opening it right away. This eliminates the bug caused
by not opening the serial device for telemetry reception.

Serial devices can now be opened only once; this eliminates errors
when trying to reflash or configure devices while receiving telemetry.

Signed-off-by: Keith Packard <keithp@keithp.com>