altosui: Replace flight status table with labels, fix resize.
There's no reason to use a table for the flight status data,
replace that with a selection of widgets instead.

Also, set all of the grid bag constraints for the various flight
status displays so that resize does something sensible.

Adds a scrollbar to the table display so that it can shrink.

Signed-off-by: Keith Packard <keithp@keithp.com>