altosui: Rewrite info table to mix with scroll pane well. Fix startup size
Using a single table for the info table means that the scroll pane
automatically picks up the table headers and shows them above the
scrollable view.

This patch also fixes the application size at startup so that no
scrollbar is required in the info table, and the window is < 800x600.

Signed-off-by: Keith Packard <keithp@keithp.com>