altosui: When fixing eeprom gps time information, make GPS data valid
Eeprom files may be missing the GPS time (due to a firmware bug). Working
around this involves finding the next valid GPS time and using that to
create a fake GPS time entry. However, that next GPS time may not
be locked or may have few sats as it is from the boost stage of the
flight. Fix this by simply forcing the fake time packet to have
4 sats and be locked.

Signed-off-by: Keith Packard <keithp@keithp.com>