altos: Switch telemetrum over to kalman filter
This changes the full telemetry stream to include kalman data instead
of the old ad-hoc flight data. It's compatible in that the packet
sizes are the same so teledongle can receive either and figure out
which it has received.

A few plotting and testing tools are added to make validating the new
code easier.

Signed-off-by: Keith Packard <keithp@keithp.com>