altos: Ignore alt error for fast->coast. Allow larger error for baro apogee.
With the fixed kalman filter, transitions across mach don't cause
bumps in the merged filter.

And, with working kalman bits, the signal for broken baro detection is
stronger and so we can allow for baro apogee detection in cases where
noise occurs close to apogee.

Bump the kalman filter to trust the baro less so that the model tracks
across mach.

Signed-off-by: Keith Packard <keithp@keithp.com>