altosui: Allow radio channel to be configured over the radio link
TeleMini/TeleNano can't be configured via USB, so we need to allow
the radio channel to be set over the radio link.

This change carefully sets the new radio channel, disables the remote
link and then sets the teledongle channel to the new value and brings
the link back up.

Signed-off-by: Keith Packard <keithp@keithp.com>