altosui: Tell serial device which frame to use for timeout dialogs
For the timeout dialog to appear, a frame must be configured for it to
appear near. This patch sends the frame from the eeprom download
functions to the serial code. That path doesn't yet work as the eeprom
download is still trying to talk to the serial device from the swing
event thread, which prevents the cancel dialog from working.

Signed-off-by: Keith Packard <keithp@keithp.com>