altosui: Handle old TeleDongle receiving kalman telemetry packets
The telemetry packets now send the kalman height/speed/accel values
instead of the ad-hoc values. If received by an old TeleDongle box,
the speed value will be of the form 0x8000abcd, which will be printed
as a 32-bit value by TeleDongle. We only want the abcd part, which is
the speed * 16. Detect this automatically and compute the correct
values for all three.

Signed-off-by: Keith Packard <keithp@keithp.com>