altos: Average height values for landing detection
Instead of using the direct output of the kalman filter and hoping
that is quiet enough to detect landing, filter that with a long
exponential decay filter and then check to make sure that doesn't
change more than 2m in 5 seconds as a trigger for landing detection.

Tested with existing telemetrum flight logs and it correctly detects
landing in all cases.

Signed-off-by: Keith Packard <keithp@keithp.com>