altos: Add 'send all baro' compile-time option
This option creates a new packet type that delivers full sensor-rate
barometer telemetry data to allow for off-line analysis of flight
algorithms using all of the data, rather than the slower rate provided
either over telemetry or stored in the eeprom file.

Define AO_SEND_ALL_BARO and this will get built in. Perhaps this could
be a run-time option...

Signed-off-by: Keith Packard <keithp@keithp.com>