altos: remove monitor disable stubs from altimeter code
Monitor mode in the ground-station boards must be disabled when the
radio is going to be used for another purpose, or the radio parameters
changed. That places monitor-mode disable calls in other parts of the
system which are shared with the altimeter code.

Elide the ao_set_monitor calls for builds which do not include any
monitoring code.

Signed-off-by: Keith Packard <keithp@keithp.com>