altos: When sharing radio DMA for AES, use it for in instead of out
We look at the out_done value, but not the in_done value; if we use
the radio DMA for out, we would have to use ao_radio_dma_done to check
for completion. This way, we can ignore that value and use the
existing ao_aes_dma_out_done value.

Signed-off-by: Keith Packard <keithp@keithp.com>