altos: SPI radio - use 1->0 for 'done' and 0->1 for 'ready'
This changes how the SPI radio protocol uses the interrupt
line. Instead of a pulse indicating operation done, this now uses a 0
value for done and a 1 value for ready. The key distinction is that
the master can tell when the slave is waiting for the next command
instead of hoping that it got done 'soon enough'.

Signed-off-by: Keith Packard <keithp@keithp.com>