altosui: Leave new_gps indication until gps value is used
During replay, AltosState may not see a new GPS value as soon as it
lands in the state field as additional records with the same timestamp
may come in after the GPS record.

Instead of resetting the new_gps indication when the new record is
created, wait until the new_gps indication is seen by the AltosState
update code and have that clear the new_gps indication.

Signed-off-by: Keith Packard <keithp@keithp.com>