altos: Time out reading packet data from cc1120 after 100ms
Sometimes the radio will give a spurious wakeup indicating that a
preamble seems to have arrived, but no packet data will appear. In
this case, abandon the packet reception and go back to waiting for a
preamble again. This releases the SPI bus for other users and also
avoids missing packets.

Signed-off-by: Keith Packard <keithp@keithp.com>