altos/micropeak: Increase boost detect to 30m
This meant increasing the data buffering as well so that we could
reliably capture the flight data back to the ground, even for slow
flights.

And, with the buffer extra large, we work backwards from the current
buffer location to find the last ground location rather than working
forwards from the first buffered location. This ensures that we don't
capture noise before boost and instead capture a nice flight curve instead.

Signed-off-by: Keith Packard <keithp@keithp.com>