altos: Allow use of internal EEPROM for config storage
This stops exposing eeprom as 'storage' and instead exposes it with a
separate eeprom API so that it can be used for config storage without
also using it for flight log storage.

The config code has been changed to allow it to either use storage for
configuration data or eeprom.

Signed-off-by: Keith Packard <keithp@keithp.com>