altosuilib: Don't match product_altusmetrum for product_basestation or product_altimeter
It's been years since we've shipped boards configured with
product_altusmetrum, but now we've repurposed that code for the flash
loader. When matching an explicit product, go ahead and also match
altusmetrum so that the flash loader will fit, but when matching
basestation or altimeter, don't as that will avoid popping up the
flight monitor UI at startup when a board is running the boot loader.

Signed-off-by: Keith Packard <keithp@keithp.com>