altos: Execute self-test on MMA655X part
This doesn't check for fine calibration, just makes sure the part is
within the broad tolerances specified by the manufacturer. The subtle fix
here was actually getting reset working -- there was a typo in the
bitpattern definition which caused the reset to fail, making it hard
to self-test the part, as self-test is only valid before the part is running.

Signed-off-by: Keith Packard <keithp@keithp.com>