altos: Flip acceleration data consistently for orientation changes
Must flip the acceleration calibration data using the same function as
flipping the actual acceleration data or the calibration will no
longer be valid after changing orientation. For the MMA655x, this
means using 4095 in both places, rather than using 0x7fff for the
calibration data changes and 4095 for the data value changes. For the
MPU6000, this means using 0 in both places.

Signed-off-by: Keith Packard <keithp@keithp.com>