doc: Replace screen captures with 'active' window ones
These are the same as the previous captures, just with the titlebar
highlighted instead of grey so that they're all consistent. These were
also done with the gimp which makes the areas outside of the window
(like outside the rounded corners) have zero alpha so that the white
background shows through.

Signed-off-by: Keith Packard <keithp@keithp.com>