altos: Put locked/unlocked GPS status in APRS comments
Replace the 'S' (which marks the field showing sats in view) with
either 'L' or 'U' to tell the user whether the GPS receiver is locked
or unlocked.

This also removes the colons in the comment field to shorten it. This
makes it fit on one line of my FT1D display.

Signed-off-by: Keith Packard <keithp@keithp.com>