altos: HAS_LED is useless; remove it
ao_flight was trying to decide whether to turn off the red LED by
checking whether HAS_LED was defined. And yet, none of the flight
firmware defines that anymore, except for easymini which defines it to
zero.

Remove all uses and defines of this value, substituting AO_LED_RED in
ao_flight.c, which has to be defined for the ao_led_off call to work.

Signed-off-by: Keith Packard <keithp@keithp.com>