altos: Clear out eeprom erase records when writing entry 0
When writing config/erase to eeprom, there's no 'erase' operation as
on-chip eeprom is writable at a byte level. As such, we can't tell
when the erase blocks get reset when the config gets written. When
this happens, erase block 0 gets written explicitly, so just use that
call to trigger explicit erasing of the data.

Signed-off-by: Keith Packard <keithp@keithp.com>