altos/telemega-v1.0: Adjust config params to make flight erase work
Flight erase records are supposed to be written after the ao_config to
eeprom in telemega. They were getting written in the middle of one of
the pyro channel config blocks. Put a bunch of space between the two
by making the config max 1024 bytes instead of 128 bytes.

Set the log erase marker to 0x55 -- eeprom comes from the factory as
0x00, so we use any value other than 0x55 to indicate 'unused' erase
slots.

Save space for more flight erase blocks; we've  got plenty.

Signed-off-by: Keith Packard <keithp@keithp.com>